requirements.txt
crispy-bootstrap4==2022.1
Django==4.2.4
django-allauth==0.54.0
django-crispy-forms==2.0
django-filter==23.2
python-decouple
djangorestframework